lab13.

+++ src/main/java/com/example/songr/infrastructure/SongRepository.java
@@ -3,6 +3,6 @@ package com.example.songr.infrastructure;
 import com.example.songr.Models.Song;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SongRepository extends CrudRepository<Song,Integer> {
+public interface SongRepository extends CrudRepository<Song, Long> {
 
 }
